Added: - WorkSpaceController ~ Controller class for WorkSpaceView inputs being passed to WorkSpaceGraph - Toolbox ~ Simple representation of a toolbox whose items can be dragged onto the WorkSpaceView - A boat load of comments that were missing from the inital sketch of the classes

+++ DataStructureViewer/src/models/WorkSpaceGraphElement.java
@@ -4,61 +4,116 @@ import javafx.geometry.Point2D;
 
 /**
  * A class which represents the data relating to a single element in a WorkSpace
- * That is all elements which are to be drawn in a work space (LinkedList Heads,
- * Arrays, Nodes etc) will all extend this class.
+ * All elements which contain the data modeling a data structure element in a
+ * work space (LinkedList Heads, Arrays, Nodes etc) will all extend this class.
+ *
  * @author Iain Workman
  */
 public abstract class WorkSpaceGraphElement {
-    
-    // Constructor
-    public WorkSpaceGraphElement(int positionX, int positionY, WorkSpaceGraph parent) {
+
+    // Constructors
+    /**
+     * Default Constructor for the abstract WorkSpaceGraphElement class. Sets
+     * the position and parent WorkSpaceGraph of this WorkSpaceGraphElement
+     *
+     * @param positionX::double ~ The X position in relative terms of the
+     * WorkSpaceGraphElement
+     * @param positionY::double ~ The Y position in relative terms of the
+     * WorkSpaceGraphElement
+     * @param parent::WorkSpaceGraph ~ The WorkSpaceGraph of which this
+     * WorkSpaceGraphElement is a node
+     */
+    public WorkSpaceGraphElement(double positionX, double positionY, WorkSpaceGraph parent) {
         position_ = new Point2D(positionX, positionY);
         parent_ = parent;
     }
-    
+
+    /**
+     * Default Constructor for the abstract WorkSpaceGraphElement class. Sets
+     * the position and parent WorkSpaceGraph of this WorkSpaceGraphElement
+     *
+     * @param position::Point2D ~ The position in relative terms of the
+     * WorkSpaceGraphElement
+     * @param parent::WorkSpaceGraph ~ The WorkSpaceGraph of which this
+     * WorkSpaceGraphElement is a node
+     */
     public WorkSpaceGraphElement(Point2D position, WorkSpaceGraph parent) {
         position_ = position;
         parent_ = parent;
     }
-    
+
     // Public Methods
+    /**
+     * @return The X position in relative terms of the WorkSpaceGraphElement
+     */
     public double getX() {
         return position_.getX();
     }
-    
+
+    /**
+     * @return The Y position in relative terms of the WorkSpaceGraphElement
+     */
     public double getY() {
         return position_.getY();
     }
-    
+
+    /**
+     * @return The position in relative terms of the WorkSpaceGraphElement
+     */
     public Point2D getPostion() {
         return position_;
     }
-    
+
+    /**
+     * Sets the X position of the WorkSpaceGraphElement in relative terms
+     *
+     * @param positionX::double ~ The X position of the WorkSpaceGraphElement
+     */
     public void setX(double positionX) {
         position_ = new Point2D(positionX, position_.getY());
         parent_.notifySubscribersOfAlter(this);
     }
-    
+
+    /**
+     * Sets the Y position of the WorkSpaceGraphElement in relative terms
+     *
+     * @param positionY::double ~ The Y position of the WorkSpaceGraphElement
+     */
     public void setY(double positionY) {
         position_ = new Point2D(position_.getX(), positionY);
         parent_.notifySubscribersOfAlter(this);
     }
-    
+
+    /**
+     * Sets the position of the WorkSpaceGraphElement in relative terms
+     *
+     * @param positionX::double ~ The X position of the WorkSpaceGraphElement
+     * @param positionY::double ~ The Y position of the WorkSpaceGraphElement
+     */
     public void setPosition(double positionX, double positionY) {
         position_ = new Point2D(positionX, positionY);
         parent_.notifySubscribersOfAlter(this);
     }
-    
+
+    /**
+     * Sets the position of the WorkSpaceGraphElement in relative terms
+     *
+     * @param position::Point2D ~ The position of the WorkSpaceGraphElement
+     */
     public void setPosition(Point2D position) {
         position_ = position;
         parent_.notifySubscribersOfAlter(this);
     }
-    
+
     // Protected Methods
+    /**
+     * @return The parent WorkSpaceGraph of which this WorkSpaceGraphElement is
+     * a node
+     */
     protected WorkSpaceGraph getParent() {
         return parent_;
     }
-    
+
     // Private Member Variables
     private Point2D position_;
     private final WorkSpaceGraph parent_;
